Refactor to shared drag utils
