+++ libs/application/templates/parental-leave/src/fields/components/Slider/index.tsx
@@ -1,5 +1,6 @@
 import React, { CSSProperties, FC, useRef, useState, useEffect } from 'react'
 import useComponentSize from '@rehooks/component-size'
+import { useDrag } from '../utils'
 
 import * as styles from './Slider.treat'
 import { Box, Text } from '@island.is/island-ui/core'
@@ -32,73 +33,9 @@ const useLatest = <T extends number>(value: T) => {
   return ref
 }
 
-interface UseDragOptions {
-  onDragStart?: () => void
-  onDragEnd?: (delta: number) => void
-  onDragMove?: (delta: number) => void
-}
-
 const roundByNum = (num: number, rounder: number) => {
   const multiplier = 1 / (rounder || 0.5)
   return Math.round(num * multiplier) / multiplier
-}
-
-const isMouseEvent = <T extends HTMLElement>(
-  event: React.MouseEvent<T, MouseEvent> | React.TouchEvent<T>,
-): event is React.MouseEvent<T, MouseEvent> => {
-  return event.nativeEvent instanceof MouseEvent
-}
-
-const useDrag = ({ onDragStart, onDragEnd, onDragMove }: UseDragOptions) => {
-  const start = useRef(0)
-
-  const handleDragMove = (event: MouseEvent | TouchEvent) => {
-    const x =
-      event instanceof MouseEvent ? event.clientX : event.targetTouches[0].pageX
-    const deltaX = x - start.current
-
-    if (onDragMove) {
-      onDragMove(deltaX)
-    }
-  }
-
-  const handleDragEnd = (event: MouseEvent | TouchEvent) => {
-    document.removeEventListener('mouseup', handleDragEnd)
-    document.removeEventListener('touchend', handleDragEnd)
-    document.removeEventListener('mousemove', handleDragMove)
-    document.removeEventListener('touchmove', handleDragMove)
-
-    const x =
-      event instanceof MouseEvent ? event.clientX : event.targetTouches[0].pageX
-    const deltaX = x - start.current
-
-    if (onDragEnd) {
-      onDragEnd(deltaX)
-    }
-  }
-
-  const handleDragStart = (
-    event:
-      | React.MouseEvent<HTMLElement, MouseEvent>
-      | React.TouchEvent<HTMLElement>,
-  ) => {
-    start.current = isMouseEvent(event)
-      ? event.clientX
-      : event.targetTouches[0].pageX
-    document.addEventListener('mousemove', handleDragMove)
-    document.addEventListener('touchmove', handleDragMove)
-    document.addEventListener('mouseup', handleDragEnd)
-    document.addEventListener('touchend', handleDragEnd)
-
-    if (onDragStart) {
-      onDragStart()
-    }
-  }
-
-  return {
-    onMouseDown: handleDragStart,
-    onTouchStart: handleDragStart,
-  }
 }
 
 interface TrackProps {
