+++ libs/application/templates/parental-leave/src/fields/components/Timeline/index.tsx
@@ -10,7 +10,7 @@ import {
   endOfMonth,
 } from 'date-fns'
 import { useWindowSize } from 'react-use'
-import { useDrag } from './utils'
+import { useDrag } from '../utils'
 
 import { Box, Icon, Text } from '@island.is/island-ui/core'
 import { theme } from '@island.is/island-ui/theme'
